% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningsChimerge.R
\name{binningsChimerge}
\alias{binningsChimerge}
\title{Title binningsChimerge}
\usage{
binningsChimerge(
  data,
  y,
  x = NULL,
  alpha = 0.1,
  limit = 0.05,
  showCols = c("vars", "bins", "count", "countRate", "nmlCount", "nmlCountCum",
    "nmlCountCumRate", "rpsCount", "rpsCountCum", "rpsCountCumRate", "rpsRate", "woe",
    "iv", "totalIv", "totalIvAdj", "binsLift", "shape", "splits", "isSpecial"),
  breaksList = NULL,
  minBinsNumInitial = NULL,
  dataLimitCondition = "all",
  responseSign = c("bad", "1", "no", "response"),
  specialValue = NULL,
  ifSpecialNeeds = 0,
  specialVarValue = NULL,
  ifSpecialVarNeeds = 0,
  binsNumLimit = 8,
  autoAdjust = 0,
  ifMonotonous = 0,
  maxIterationLimit = Inf,
  supplyCl = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{y}{Response value label column name.}

\item{x}{Variable value label column name.}

\item{alpha}{Chi-square test threshold.}

\item{limit}{The smallest proportion of bins.}

\item{showCols}{Column of results,The full column name is c('vars', 'bins', 'count', 'countRate', 'nmlCount', 'nmlCountCum', 'nmlCountCumRate', 'rpsCount', 'rpsCountCum', 'rpsCountCumRate', 'rpsRate', 'woe', 'iv', 'totalIv', 'totalIvAdj', 'binsLift', 'shape', 'splits', 'isSpecial')}

\item{breaksList}{Custom binning value, must be in list form.}

\item{minBinsNumInitial}{The minimum number of bins, the priority is above the limit parameter.}

\item{dataLimitCondition}{When using the limit parameter, how to choose the number of data sets, "all" or "DelNa" or "DelNaSpe" is optional.}

\item{responseSign}{Response value label name.Default is c('bad','1','no','response').}

\item{specialValue}{Values that need to be binned separately, which are globally valid for the data set.}

\item{ifSpecialNeeds}{Whether the global special value remains in the data when binning.}

\item{specialVarValue}{The value that needs to be binned separately, valid for the specified variable, must be a list.}

\item{ifSpecialVarNeeds}{Whether the special value of the variable dimension is included in the calculation when binning.}

\item{binsNumLimit}{The minimum number of bins is limited, does not include missing values, special values.}

\item{autoAdjust}{Whether to adjust automatically, automatic adjustment can adjust the final binning result to monotonous response rate or U-shaped.}

\item{ifMonotonous}{Whether the binning is monotonous, the final binning result must be monotonous and optimal, which is very effective for some situations, and it is not recommended to be used in conjunction with autoAdjust.}

\item{maxIterationLimit}{Regarding the maximum number of iterations for a single variable, some computers cannot withstand the limitation of too many iterations. In most cases, it is not necessary to set it.}

\item{supplyCl}{The number of cores in the parallel backend, considering that most of the analysis will be performed on the server, in order to avoid affecting others, half of the core number is used by default.}
}
\value{
A binned list of all variables.
}
\description{
Chi-square optimal binning function,A bottom-up approach to merging.
}
\examples{
library(data.table)

data("germancredit", package="scorecard")

data <- germancredit

##  ALL

binsChi <- binningsChimerge(
 data = data
  , y = 'creditability'
)

binsChiDt <- rbindlist(binsChi)

##  One

binsChiOne <- binningsChimerge(
  data = data
  , y = 'creditability'
  , x = 'duration.in.month'
)

}
