% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningsTraverse.R
\name{binningsTraverse}
\alias{binningsTraverse}
\title{Title binningsTraverse}
\usage{
binningsTraverse(
  data,
  y,
  x = NULL,
  limit = 0.05,
  showCols = c("vars", "bins", "count", "countRate", "nmlCount", "nmlCountCum",
    "nmlCountCumRate", "rpsCount", "rpsCountCum", "rpsCountCumRate", "rpsRate", "woe",
    "iv", "totalIv", "totalIvAdj", "binsLift", "shape", "splits", "isSpecial"),
  responseSign = c("bad", "1", "no", "坏", "response"),
  supplyCl = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{y}{Response value label column name.}

\item{x}{Variable value label column name.}

\item{limit}{The smallest proportion of bins.}

\item{showCols}{Column of results,The full column name is c('vars', 'bins', 'count', 'countRate', 'nmlCount', 'nmlCountCum', 'nmlCountCumRate', 'rpsCount', 'rpsCountCum', 'rpsCountCumRate', 'rpsRate', 'woe', 'iv', 'totalIv', 'totalIvAdj', 'binsLift', 'shape', 'splits', 'isSpecial').}

\item{responseSign}{Response value label name.Default is c('bad','1','no','response').}

\item{supplyCl}{The number of cores in the parallel backend, considering that most of the analysis will be performed on the server, in order to avoid affecting others, half of the core number is used by default.}
}
\value{
A binned list of all variables.
}
\description{
Iterate over all values of a variable for binning operation.
}
\examples{

library(data.table)

data("germancredit", package="scorecard")

data <- germancredit

##  ALL

binsTraverse <- binningsTraverse(
  data = data
  , y = 'creditability'
)

binsTraverseDt <- rbindlist(binsTraverse)

##  One

binsTraverseOne <- binningsTraverse(
  data = data
  , y = 'creditability'
  , x = 'housing'
)


}
