% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsBestKS.R
\name{binsBestKS}
\alias{binsBestKS}
\title{Title}
\usage{
binsBestKS(
  data,
  y,
  x,
  ksLimit = 0.03,
  initialBins = NULL,
  limit = 0.05,
  breaksList = NULL,
  minBinsNumInitial = NULL,
  dataLimitCondition = "all",
  responseSign = c("bad", "1", "no", "response"),
  specialValue = NULL,
  ifSpecialNeeds = 0,
  specialVarValue = NULL,
  ifSpecialVarNeeds = 0,
  binsNumLimit = 8,
  autoAdjust = 0,
  ifMonotonous = 0,
  maxIterationLimit = Inf
)
}
\arguments{
\item{maxIterationLimit}{}
}
\value{

}
\description{
Title
}
\examples{

}
