% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsCluster.R
\name{binsCluster}
\alias{binsCluster}
\title{Title}
\usage{
binsCluster(
  data,
  x,
  nbins = 5,
  method = "K-Medoids",
  specialValue = NULL,
  ifSpecialNeeds = 0,
  specialVarValue = NULL,
  ifSpecialVarNeeds = 0
)
}
\arguments{
\item{ifSpecialVarNeeds}{}
}
\value{

}
\description{
Title
}
