% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsEqual.R
\name{binsEqual}
\alias{binsEqual}
\title{Title  binsEqual}
\usage{
binsEqual(
  data,
  x,
  nbins = 5,
  specialValue = NULL,
  ifSpecialNeeds = 0,
  specialVarValue = NULL,
  ifSpecialVarNeeds = 0
)
}
\arguments{
\item{ifSpecialVarNeeds}{}
}
\value{
Returns the binned values of equal frequency bins.
}
\description{
Equal frequency binning
}
