% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsWidth.R
\name{binsWidth}
\alias{binsWidth}
\title{Title}
\usage{
binsWidth(
  data,
  x,
  nbins = 5,
  specialValue = NULL,
  ifSpecialNeeds = 0,
  specialVarValue = NULL,
  ifSpecialVarNeeds = 0
)
}
\arguments{
\item{ifSpecialVarNeeds}{}
}
\value{

}
\description{
Title
}
