% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiMergeCha.R
\name{chiMergeCha}
\alias{chiMergeCha}
\title{Title}
\usage{
chiMergeCha(
  data,
  y,
  x,
  alpha = 0.1,
  limit = 0.05,
  breaksList = NULL,
  minBinsNumInitial = NULL,
  dataLimitCondition = "all",
  responseSign = c("bad", "1", "no", "response"),
  specialValue = NULL,
  ifSpecialNeeds = 0,
  specialVarValue = NULL,
  ifSpecialVarNeeds = 0,
  binsNumLimit = 8,
  maxIterationLimit = Inf
)
}
\arguments{
\item{maxIterationLimit}{}
}
\value{

}
\description{
Title
}
