% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeData.R
\name{describeData}
\alias{describeData}
\title{Title describeData}
\usage{
describeData(data, nBins = 8)
}
\arguments{
\item{data}{A data frame.}

\item{nBins}{Number of bins for small graphics.}
}
\value{
Basic data statistics
}
\description{
Basic data statistics.
}
\examples{

library(data.table)

data("germancredit", package="scorecard")

data <- germancredit

baseSta <- describeData(data)
baseSta

}
