% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskTableNum.R
\name{riskTableNum}
\alias{riskTableNum}
\title{Title}
\usage{
riskTableNum(
  data,
  y,
  x,
  showCols = c("vars", "bins", "count", "countRate", "nmlCount", "nmlCountCum",
    "nmlCountCumRate", "rpsCount", "rpsCountCum", "rpsCountCumRate", "rpsRate", "woe",
    "iv", "totalIv", "totalIvAdj", "binsLift", "shape", "splits", "isSpecial"),
  specialValue = NULL,
  specialVarValue = NULL,
  responseSign = c("bad", "1", "no", "response"),
  splitValue
)
}
\arguments{
\item{splitValue}{}
}
\value{

}
\description{
Title
}
