% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCard.R
\name{scoreCard}
\alias{scoreCard}
\title{Title scoreCard}
\usage{
scoreCard(bins, logitModel, points0 = 600, odds = 1/19, pdo = 50, scipen = 0)
}
\arguments{
\item{bins}{Variable binning, which can be a list or a data frame.}

\item{logitModel}{Results of the base glm logistic regression model}

\item{points0}{Target points, default 600.}

\item{odds}{Target odds, default 1/19. Odds = p/(1-p).}

\item{pdo}{Points to Double the Odds, default 50.}

\item{scipen}{Fractions with decimal places.}
}
\value{
Scorecard conversion scores for all observations.
}
\description{
Model scores for each bin interval.
}
\examples{
library(data.table)

data("germancredit", package="scorecard")

data <- germancredit

##  Convert Y labels

data <- transformResponse(data,y = 'creditability')

dataList <- splitDt(data = data,ratio = c(0.7,0.3),seed = 520)

labelList <- lapply(dataList, function(x) x$y)

##  bins

binsChi <- binningsChimerge(
  data = data
  , y = 'y'
)

binsChiDt <- rbindlist(binsChi)

## woe

woeData <- woeTrans(data,y = 'y',bins = binsChi)

woeDataList <- lapply(dataList,function(x) woeTrans(x,y = 'y',bins = binsChi))

## glm

fit = glm(
  y ~ .
  , family = binomial()
  , data = woeDataList$train
)

summary(fit)

## stepwise by aic

fitStepAic <- step(
  fit
  , direction = 'both'
  , k = 2
  , trace = FALSE
)

summary(fitStepAic)

# ## stepwise by bic
#
# fitStepBic <- step(
#   fit
#   , direction = 'both'
#   , k = log(nrow(woeDataList$train)))
#   , trace = FALSE
# )
#
# summary(fitStepBic)

## pred

predList <- lapply(woeDataList, function(x) predict(fitStepAic, x, type='response'))

## performance

modelPerformance <- modelEffect(
  pred = predList
  , label = labelList
  , ord = "desc"
)

modelPerformance

##  score and scorecard

### Only output the final model score

modelScoreRes <- lapply(predList, function(x) modelScore(x))

### scorecard

sc <- scoreCard(
  bins = binsChi
  , logitModel = fitStepAic
)

scDt <- rbindlist(sc, fill = T)

riskForm(
 pred = modelScoreRes
 , label = labelList
 , ord = 'asc'
)


}
