% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDt.R
\name{splitDt}
\alias{splitDt}
\title{Title}
\usage{
splitDt(
  data,
  ratio = c(0.7, 0.3),
  dataNames = c("train", "test"),
  ifKeepI = FALSE,
  seed = 520,
  kFold = NULL,
  kFoldRate = 0.8
)
}
\arguments{
\item{data}{A data frame.}

\item{ratio}{The ratio of the two datasets before and after.}

\item{dataNames}{The names of the two datasets before and after.}

\item{ifKeepI}{Whether to keep the row count index of the original row.}

\item{seed}{Random seed.}

\item{kFold}{K-fold cross split, the priority is the highest, if there is a value, the previous parameters are ignored.}

\item{kFoldRate}{K-fold cross ratio.}
}
\value{
A list containing the training set, test set, or k-fold crossover data.
}
\description{
Title
}
\examples{

library(data.table)

data("germancredit", package="scorecard")

data <- germancredit

useData <- splitDt(data)

useDataKfold <- splitDt(data,kFold = 5,kFoldRate = 0.75)

}
