% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformResponse.R
\name{transformResponse}
\alias{transformResponse}
\title{Title transformResponse}
\usage{
transformResponse(
  data,
  y,
  responseSign = c("bad", "1", "no", "response"),
  ifKeepOriY = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{y}{Response value label column name.}

\item{responseSign}{Response value label name.Default is c('bad','1','no','response').}

\item{ifKeepOriY}{Whether the original Y label is saved, the default is not to save.}
}
\value{
Raw data add a column of Y label.
}
\description{
Y label replacement on raw data.
}
\examples{
data("germancredit", package="scorecard")

data <- germancredit

newData <- transformResponse(data,'creditability')



}
