% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woePlotGG.R
\name{woePlotGG}
\alias{woePlotGG}
\title{Title woePlotGG}
\usage{
woePlotGG(bins)
}
\arguments{
\item{bins}{}
}
\value{
ggplot2 image for binning of variables
}
\description{
Visualize variable binning via ggplot2
}
\examples{
library(data.table)

data("germancredit", package="scorecard")

data <- germancredit

binsChi <- binningsChimerge(
  data = data
  , y = 'creditability'
)


woePlotGG(binsChi)

woePlotGG(binsChi$duration.in.month)


}
