% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woeTrans.R
\name{woeTrans}
\alias{woeTrans}
\title{Title}
\usage{
woeTrans(data, y = NULL, x = NULL, bins, repClass = "woe", supplyCl = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{y}{Response value label column name.}

\item{x}{Variable value label column name.}

\item{bins}{Variable binning, which can be a list or a data frame.}

\item{repClass}{It can only be converted to woe for the time being, and bins will be added later.}

\item{supplyCl}{The number of cores in the parallel backend, considering that most of the analysis will be performed on the server, in order to avoid affecting others, half of the core number is used by default.}
}
\value{
The data frame converted by woe encoding.
}
\description{
Title
}
\examples{
library(data.table)

data("germancredit", package="scorecard")

data <- germancredit

##  ALL

binsChi <- binningsChimerge(
  data = data
  , y = 'creditability'
)

woeData <- woeTrans(data = germancredit,y = 'creditability',bins = binsChi)


}
